Move allElements to LibraryReader
Fixes #229

LibraryReader is a nicer place for this utility and will give us a
surface to attach more functionality than scattering top-level methods.
