+++ lib/source_gen.dart
@@ -12,4 +12,3 @@ export 'src/library.dart' show LibraryReader;
 export 'src/revive.dart' show Revivable;
 export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker;
-export 'src/utils.dart' show allElements;
